Added Mission Feeder

+++ src/main/java/com/mycompany/gs1_prototipo1/model/Coordinate.java
@@ -1,19 +1,19 @@
 package com.mycompany.gs1_prototipo1.model;
 
 public class Coordinate {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
 
     public Coordinate(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
